Move `updateVersions()` to tasks.
