+++ packages/ckeditor5-dev-release-tools/lib/index.js
@@ -14,7 +14,7 @@ const updateCKEditor5Dependencies = require( './tasks/updateckeditor5dependencie
 const updateDependenciesVersions = require( './utils/updatedependenciesversions' );
 const { getLastFromChangelog, getCurrent, getLastTagFromGit } = require( './utils/versions' );
 const { getChangesForVersion, getChangelog, saveChangelog } = require( './utils/changelog' );
-const updateVersions = require( './release/updateversions' );
+const updateVersions = require( './tasks/updateversions' );
 
 module.exports = {
 	releaseSubRepositories,
